Hide markers during SessionState Loading
UnHide is with a delay of 1000 ms, because after the session switch from "Loading" to "Race" the position data for the markers is not completely available, mainly for the human driver.

The UnHide Timer uses the sensorLayer_UpdateDelta to measure the time, because this is the time from worker run to worker run and the addition of it grows with the real time. In result there is no need of a new Date object.