draw call in update function disabled
The draw function called itself by the GPSSensor in init_map. The draw only works correct if there is data  available at this moment -> https://developers.google.com/maps/documentation/javascript/customoverlays

But the data is initialized by the worker and the first worker run is finished after the init_map.
In result additionally filled the aSensorData with init data.