Fix for Issue #99
RefreshDriverList is now called before the export writes the data in every export trigger.
This is needed, bcause the table is not updated, if it is collapsed.
SessionState and SessionStage checks for these triggers moved to a later position, because at the old one there is not all data available for RefreshDriverList.
If the drivertable is unfolded, RefreshDriverList is called two times at a moment where a trigger activates an export. One time for the trigger and second time the regular call.